+++ src/v2/modules/invoice-accounts/routes.js
@@ -51,28 +51,6 @@ exports.createInvoiceAccount = {
   }
 };
 
-exports.postInvoiceAccountAddress = {
-  method: 'POST',
-  path: '/crm/2.0/invoice-accounts/{invoiceAccountId}/addresses',
-  handler: (request, h) => entityHandlers.createEntity(request, h, 'invoiceAccountAddress'),
-  options: {
-    description: 'Adds an address to an invoice account entity',
-    validate: {
-      params: {
-        invoiceAccountId: validators.GUID
-      },
-      payload: {
-        addressId: validators.GUID,
-        startDate: validators.START_DATE,
-        endDate: validators.END_DATE,
-        agentCompanyId: validators.GUID.allow(null).default(null),
-        contactId: validators.GUID.allow(null).default(null),
-        isTest: validators.TEST_FLAG
-      }
-    }
-  }
-};
-
 exports.deleteInvoiceAccount = {
   method: 'DELETE',
   path: '/crm/2.0/invoice-accounts/{invoiceAccountId}',
@@ -82,21 +60,6 @@ exports.deleteInvoiceAccount = {
     validate: {
       params: {
         invoiceAccountId: validators.GUID
-      }
-    }
-  }
-};
-
-exports.deleteInvoiceAccountAddress = {
-  method: 'DELETE',
-  path: '/crm/2.0/invoice-accounts/{invoiceAccountId}/addresses/{invoiceAccountAddressId}',
-  handler: (request, h) => entityHandlers.deleteEntity(request, h, 'invoiceAccountAddress'),
-  options: {
-    description: 'Delete an invoice account address entity by id',
-    validate: {
-      params: {
-        invoiceAccountId: validators.GUID,
-        invoiceAccountAddressId: validators.GUID
       }
     }
   }
